Remove myself from `CODE_OWNERS.txt`.
It's been years since I've been active on LLVM changes and code reviews,
and other folks are much more active and better options here. I'm not
explicitly trying to add specific coverage as I think the list of
covered areas is probably not even an accurate reflection of what is
needed and how best to identify good coverage. I'll leave it to the
active contributors to figure out what if any code owners make sense to
add to backfill here. Also, still happy to discuss any of the parts of
LLVM and try to provide suggestions or clarifications if useful to
folks. =]